% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn3d.R
\name{plot_venn3d}
\alias{plot_venn3d}
\title{Plot Venn diagram for 3-dimensional data.}
\usage{
plot_venn3d(
  x,
  labels = c("A", "B", "C"),
  Colors = c("red", "yellow", "orange", "lightblue", "purple", "green", "grey"),
  Title = NULL,
  num.size = 18,
  shrink = 1,
  rot = 0,
  printvals = TRUE
)
}
\arguments{
\item{x}{a numeric vector of length 7, with names c("001","010","011","100","101","110","111") in suitable order.}

\item{labels}{a character vector of length 2, providing names for the 2 groups/dimensions.}

\item{Colors}{a vector of color names for the backgrounds of each part of the diagram.}

\item{Title}{optional: a character vector of length 1, specifying title for the whole plot.}

\item{num.size}{sets the font size for the numbers inside the Venn diagram.}

\item{shrink}{a numeric value, specifying zooming effect of the plot, defaults to 1.}

\item{rot}{a numeric value, specifying the number of degrees to rotate the graph.}

\item{printvals}{boolean value indicating whether to print the values of the groups. default=TRUE}
}
\description{
Given a vector of 7 values, which describe 3-dimensional data, it plots up Venn diagram, i.e. 'crossing circles'. The user is able to specify values, labels for each circle-group and colors.
}
\examples{
y <- c(37,29,6,232,121,77,25)
names(y) <- c("001","010","011","100","101","110","111")
labels <- c("A","B","C")
plot.new()
plot_venn3d(y, labels, Colors=rainbow(7), Title = "This is an example of a 3D Venn plot")

}
