% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVenn.R
\name{plotVenn}
\alias{plotVenn}
\title{plotVenn}
\usage{
plotVenn(
  df,
  title = "",
  not_in_set_name = "None",
  shape = "circle",
  input = "disjoint",
  control = list(),
  random_seed = 1,
  legend = TRUE,
  fills = TRUE,
  edges = TRUE
)
}
\arguments{
\item{df}{Input data frame, one row per element, one column per set. 1 if that element is in that set, 0 otherwise. Column names are set names.}

\item{title}{Title for the plot}

\item{not_in_set_name}{Name for the set of elements not included in the sets provided, default "None". If not_in_set_name = NULL then don't plot these elements.}

\item{shape}{shape can be either "circle" (default) or "ellipse". "ellipse" will be able to fit area to size of set in more cases.}

\item{input}{either "disjoint" (default) or "union". See eulerr::euler().}

\item{control}{See eulerr::euler().}

\item{random_seed}{Set the random seed to fix the orientation of the plot to be reproducible. Try a few numbers till you get one you like the look of.}

\item{legend}{TRUE Should the plot have a legend? TRUE (default) or FALSE}
}
\value{
returns a base R plot of a Venn diagram.
}
\description{
Wrapper for eulerr::euler() to enable quick and painless venn diagrams.
}
\examples{
cars <- mtcars \%>\%
 transmute(`High Efficiency` = mpg>20,
           `More Cylinders` = cyl>=6)

plotVenn(cars, title = "Overlap between more cylinders\n and high efficiency in mtcars")

}
