% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_circles.R
\name{resize_circles}
\alias{resize_circles}
\title{Match the areas of the 2 circles and their overlap to the input values for X, Y and the intersection of X and Y.}
\usage{
resize_circles(x_only, y_only, overlap, standardDistance = sqrt(0.5))
}
\arguments{
\item{x_only}{a numeric value representing the relative size of X excluding its intersection with Y.}

\item{y_only}{a numeric value representing the relaitive size of Y excluding its intersection with X.}

\item{overlap}{a numeric value representing the relative size of the intersection of X and Y.}

\item{standardDistance}{a numeric value specifying the fixed center-to-center distance to which the radii are matched. The default value is that used by plotVenn2d}
}
\description{
Calculates the radii of the two circles relative to a fixed center-to-center distance. The areas of the circles and their overlap will be proportionate the three input values of X alone, Y alone and the intersection of X and Y.
}
\examples{
 plot.new()
plot_venn2d(rep("",3), radius=resize_circles(20,.5, 3), Title=NULL, 
          resizePlot=0.7, labels=c("",""))
          
}
